Feature: Batch upload.
Unfortunatel, the stable version of apache's httpclient doesn't support preemptive authentication, for security reasons, so I had to upgrade to 4.1 alpha. This feature is only being used for batch upload, so if we want to downgrade in the future look at this commit.


Also, added Apache license notifications to files.